# swagger.json
{
	"swagger": "2.0",
	"info": {
		"description": "Api Documentation",
		"version": "1.0",
		"title": "Api Documentation",
		"termsOfService": "urn:tos",
		"contact": {
			
		},
		"license": {
			"name": "Apache 2.0",
			"url": "http://www.apache.org/licenses/LICENSE-2.0"
		}
	},
	"host": "localhost:9090",
	"basePath": "/",
	"tags": [
		{
			"name": "contact-controller",
			"description": "Contact Controller"
		}
	],
	"paths": {
		"/api/contact": {
			"post": {
				"tags": [
					"contact-controller"
				],
				"summary": "contact",
				"operationId": "contactUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "contact",
						"description": "contact",
						"required": true,
						"schema": {
							"$ref": "#/definitions/Contact"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/contact/{cid}": {
			"get": {
				"tags": [
					"contact-controller"
				],
				"summary": "getContact",
				"operationId": "getContactUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cid",
						"in": "path",
						"description": "cid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/Contact"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"delete": {
				"tags": [
					"contact-controller"
				],
				"summary": "deleteContact",
				"operationId": "deleteContactUsingDELETE",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "cid",
						"in": "path",
						"description": "cid",
						"required": true,
						"type": "integer",
						"format": "int32"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		},
		"/api/contacts": {
			"get": {
				"tags": [
					"contact-controller"
				],
				"summary": "getAllContacts",
				"operationId": "getAllContactsUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/Contact"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		}
	},
	"definitions": {
		"Contact": {
			"type": "object",
			"properties": {
				"activeSw": {
					"type": "string"
				},
				"cid": {
					"type": "integer",
					"format": "int32"
				},
				"email": {
					"type": "string"
				},
				"name": {
					"type": "string"
				},
				"phno": {
					"type": "string"
				}
			},
			"title": "Contact"
		}
	}
}